fix(ci): split heavy EpubHarness progress-sync tests; 90 s GoldenTrace @After sleep
Two root causes behind the latest phone-harness failures:

1. EpubHarnessTest grew from 2 to 4 tests after the 13-test restore
   (progressSyncUsesCorrectEndpoint + progressSyncSendsEpubCfiNotJson
   were added back). Each progressSync test holds the Readium WebView
   active for ~40 s waiting for a progress-sync round-trip, leaving far
   more native memory behind than a quick reader-open test. Three
   consecutive heavy opens exhausted the API-25 renderer:
   • opensEpubViaSeriesNavigation → 15 s series-nav UI timeout (emulator
     sluggish under memory pressure)
   • progressSyncUsesCorrectEndpoint → 90 s WebView-open stall

   Fix: move progressSyncUsesCorrectEndpoint + progressSyncSendsEpubCfiNotJson
   to a new EpubProgressHarnessTest class. It auto-routes to shard 3
   (REST catch-all) where 200+ feature tests provide a long GC gap before
   those 2 heavy WebViews open. EpubHarnessTest is back to 2 tests
   (1 light + 1 heavy), matching the green-run baseline.

2. In shard 4 (GoldenTrace → NavSnap → Search), GoldenTrace's single
   WebView session is uniquely heavy: it logs into ABS, opens the EPUB
   reader, and keeps Readium active for ~40 s while syncing progress.
   Even in a fresh renderer after the shard-3 sleep, GoldenTrace's
   accumulated native memory prevented NavSnap cross-chapter (the very
   next WebView open) from starting — stalling for 90 s:
   com.riffle.app.harness.NavigationSnapHarnessTest
     > crossChapterJump_landsSnapped_noMovementAfterCoverLifts FAILED
       ComposeTimeoutException after 90000 ms

   Fix: add a 90 s Thread.sleep in GoldenTraceHarnessTest @After, matching
   the inter-shard sleep. After GoldenTrace's test closes, gc() runs,
   then 90 s of OS eviction under 1.5 GB RAM pressure clears ashmem pages
   before NavSnap opens its reader.

Also: raise harness-test-phone timeout-minutes from 40 → 45 (GoldenTrace's
90 s @After sleep adds ~1.5 min to shard 4's wall-clock time).
